scripts/requirements.txt
pytube==12.1.2
ffmpeg-python==0.2.0
requests==2.28.2